requirement.txt
absl-py==2.3.1
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
albumentations==0.4.3
altair==5.4.1
antlr4-python3-runtime==4.8
async-timeout==5.0.1
attrs==25.3.0
blinker==1.8.2
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
contourpy==1.1.1
cycler==0.12.1
einops==0.3.0
filelock==3.16.1
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2025.3.0
ftfy==6.2.3
future==1.0.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.40.3
google-auth-oauthlib==1.0.0
grpcio==1.70.0
hf-xet==1.1.8
huggingface-hub==0.34.4
idna==3.10
imageio==2.9.0
imageio-ffmpeg==0.4.2
imgaug==0.2.6
importlib_metadata==8.5.0
importlib_resources==6.4.5
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
kornia==0.6.4
-e git+https://github.com/OT20031003/Latent-Diffusion.git@7338d9d7d71b27727beffb943a6a05e226d9ae3d#egg=latent_diffusion
lazy_loader==0.4
lightning-utilities==0.11.9
lpips==0.1.4
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mdurl==0.1.2
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186064646/work
mkl-service==2.4.0
mpmath==1.3.0
multidict==6.1.0
narwhals==1.42.1
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
oauthlib==3.3.1
omegaconf==2.1.1
opencv-python==4.1.2.30
opencv-python-headless==4.12.0.88
packaging==24.2
pandas==2.0.3
pillow @ file:///croot/pillow_1721059439630/work
pkgutil_resolve_name==1.3.10
propcache==0.2.0
protobuf==5.29.5
pudb==2019.2
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydeck==0.9.1
pyDeprecate==0.3.1
Pygments==2.19.2
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytorch-lightning==1.4.2
pytz==2025.2
PyWavelets==1.4.1
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.20.1
rsa==4.9.1
sacremoses==0.1.1
scikit-image==0.20.0
scipy==1.9.1
seaborn==0.13.2
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.2
streamlit==1.40.1
sympy==1.13.3
tenacity==9.0.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
test_tube==0.7.5
tifffile==2023.7.10
tokenizers==0.12.1
toml==0.10.2
torch==2.4.1
torch-fidelity==0.3.0
torchmetrics==0.7.0
torchvision==0.8.1
tornado==6.4.2
tqdm==4.67.1
transformers==4.19.2
triton==3.0.0
typing_extensions @ file:///croot/typing_extensions_1715268824938/work
tzdata==2025.2
urllib3==2.2.3
urwid==2.6.16
watchdog==4.0.2
wcwidth==0.2.13
Werkzeug==3.0.6
yarl==1.15.2
zipp==3.20.2
